feat: update amplify delete to also delete all PS values for the project (renaming and abstracting)
