+++ packages/amplify-provider-awscloudformation/src/index.ts
@@ -50,8 +50,8 @@ import { LocationService } from './aws-utils/aws-location-service';
 import { hashDirectory } from './upload-appsync-files';
 import { prePushCfnTemplateModifier } from './pre-push-cfn-processor/pre-push-cfn-modifier';
 import { getApiKeyConfig } from './utils/api-key-helpers';
-import { getEnvParametersDeleteHandler } from './utils/delete-from-parameter-store';
-export { getEnvParametersDeleteHandler } from './utils/delete-from-parameter-store';
+import { getEnvParametersDeleteHandler } from './utils/delete-from-service';
+export { getEnvParametersDeleteHandler } from './utils/delete-from-service';
 
 function init(context) {
   return initializer.run(context);
